Fork update

+++ www/details.php
@@ -4,6 +4,9 @@ require_once('object_detail.inc');
 require_once('page_data.inc');
 require_once('waterfall.inc');
 
+// Prevent the details page from running out of control.
+set_time_limit(30);
+
 require_once __DIR__ . '/include/TestInfo.php';
 require_once __DIR__ . '/include/TestRunResults.php';
 require_once __DIR__ . '/include/RunResultHtmlTable.php';
@@ -14,21 +17,18 @@ require_once __DIR__ . '/include/RequestDetailsHtmlSnippet.php';
 require_once __DIR__ . '/include/RequestHeadersHtmlSnippet.php';
 require_once __DIR__ . '/include/AccordionHtmlHelper.php';
 
-$options = null;
-if (array_key_exists('end', $_REQUEST))
-    $options = array('end' => $_REQUEST['end']);
 $testInfo = TestInfo::fromFiles($testPath);
-$testRunResults = TestRunResults::fromFiles($testInfo, $run, $cached, null, $options);
-$data = loadPageRunData($testPath, $run, $cached, $options, $test['testinfo']);
+$testRunResults = TestRunResults::fromFiles($testInfo, $run, $cached, null);
+$data = loadPageRunData($testPath, $run, $cached, $test['testinfo']);
 $isMultistep = $testRunResults->countSteps() > 1;
 
-$page_keywords = array('Performance Test','Details','Webpagetest','Website Speed Test','Page Speed');
+$page_keywords = array('Performance Test','Details','WebPageTest','Website Speed Test','Page Speed');
 $page_description = "Website performance test details$testLabel";
 ?>
 <!DOCTYPE html>
 <html>
     <head>
-        <title>WebPagetest Test Details<?php echo $testLabel; ?></title>
+        <title>WebPageTest Test Details<?php echo $testLabel; ?></title>
         <?php $gaTemplate = 'Details'; include ('head.inc'); ?>
         <style type="text/css">
         div.bar {
@@ -138,7 +138,7 @@ $page_description = "Website performance test details$testLabel";
                         if( is_file("$testPath/{$run}{$cachedText}_dynaTrace.dtas") )
                         {
                             echo "<br><a href=\"/$testPath/{$run}{$cachedText}_dynaTrace.dtas\">Download dynaTrace Session</a>";
-                            echo ' (<a href="http://ajax.dynatrace.com/pages/" target="_blank">get dynaTrace</a>)';
+                            echo ' (<a href="https://www.dynatrace.com/topics/ajax-edition/" target="_blank">get dynaTrace</a>)';
                         }
                         if( is_file("$testPath/{$run}{$cachedText}_bodies.zip") )
                             echo "<br><a href=\"/$testPath/{$run}{$cachedText}_bodies.zip\">Download Response Bodies</a>";
@@ -158,7 +158,7 @@ $page_description = "Website performance test details$testLabel";
                 {
                     require_once('google/google_lib.inc');
                 ?>
-                    <h2>Csi Metrics</h2>
+                    <h2>CSI Metrics</h2>
                             <table id="tableCustomMetrics" class="pretty" align="center" border="1" cellpadding="10" cellspacing="0">
                                <tr>
                             <?php
@@ -170,13 +170,14 @@ $page_description = "Website performance test details$testLabel";
                                 echo "</tr>\n";
                                 foreach ($testRunResults->getStepResults() as $stepResult) {
                                     echo "<tr>\n";
-                                    $params = ParseCsiInfoForStep($stepResult->createTestPaths(), true);
+                                    if (GetSetting('enable_csi'))
+                                      $params = ParseCsiInfoForStep($stepResult->createTestPaths(), true);
                                     if ($isMultistep) {
                                         echo '<td class="even" valign="middle">' . $stepResult->readableIdentifier() . '</td>';
                                     }
                                     foreach ( $test['testinfo']['extract_csi'] as $csi_param )
                                     {
-                                        if( array_key_exists($csi_param, $params) )
+                                        if( isset($params) && array_key_exists($csi_param, $params) )
                                         {
                                             echo '<td class="even" valign="middle">' . $params[$csi_param] . '</td>';
                                         }
@@ -216,6 +217,7 @@ $page_description = "Website performance test details$testLabel";
                         echo "<td><a href='#request_headers_$stepSuffix'>Request Headers</a></td>";
                         echo "<td><a href='" . $urlGenerator->stepDetailPage("customWaterfall", "width=930") . "'>Customize Waterfall</a></td>";
                         echo "<td><a href='" . $urlGenerator->stepDetailPage("pageimages") . "'>All Images</a></td>";
+                        echo "<td><a href='" . $urlGenerator->stepDetailPage("http2_dependencies") . "'>HTTP/2 Dependency Graph</a></td>";
                         echo "</tr>";
                     }
                     echo "</table>\n<br>\n";
@@ -266,7 +268,7 @@ $page_description = "Website performance test details$testLabel";
                 <br>
                 <?php
                     echo '';
-                    if (array_key_exists('testinfo', $test) && array_key_exists('testerDNS', $test['testinfo']) && strlen($test['testinfo']['testerDNS']))
+                    if (isset($test) && is_array($test) && isset($test['testinfo']['testerDNS']))
                         echo "<p>Test Machine DNS Server(s): {$test['testinfo']['testerDNS']}</p>\n";
 
                     if ($isMultistep) {
@@ -359,9 +361,9 @@ $page_description = "Website performance test details$testLabel";
                 stepNum = 1;
                 doExpandAll = true;
             }
-            if (stepNum <= 0) {
-                return;
-            }
+           if (stepNum <= 0) {
+           return;
+           }
             var expand = function() {
                 var scrollToNode = $(window.location.hash);
                 if (doExpandAll) {
@@ -374,7 +376,9 @@ $page_description = "Website performance test details$testLabel";
             };
             var slide_opener = $("#request_headers_step" + stepNum);
             if (slide_opener.length) {
-              accordionHandler.toggleAccordion(slide_opener, true, expand);
+                <?php if ($isMultistep) { ?>
+                accordionHandler.toggleAccordion(slide_opener, true, expand);
+                <?php } ?>
             } else {
                 expand();
             }
@@ -385,7 +389,9 @@ $page_description = "Website performance test details$testLabel";
             if (!hash) {
                 var defaultAccordion = $("#waterfall_view_step1");
                 if (defaultAccordion.length) {
-                  accordionHandler.toggleAccordion(defaultAccordion);
+                    <?php if ($isMultistep) { ?>
+                    accordionHandler.toggleAccordion(defaultAccordion);
+                    <?php } ?>
                 }
                 return;
             }
@@ -393,7 +399,9 @@ $page_description = "Website performance test details$testLabel";
                 hash.startsWith("#connection_view_step") ||
                 hash.startsWith("#request_details_step") ||
                 hash.startsWith("#request_headers_step")) {
+              <?php if ($isMultistep) { ?>
               accordionHandler.handleHash();
+              <?php } ?>
             }
             handleRequestHash();
         }
@@ -417,4 +425,3 @@ $page_description = "Website performance test details$testLabel";
         </script>
     </body>
 </html>
-
